working upload function and cleaned all directory
