--- test_results/suggested_response_scripts.csv
@@ -1,25 +1,0 @@
-Type,Topic,Original,Suggested_Response
-Calm,Payment,"I was promised a discount on my service, but the invoice doesn’t reflect it. What’s going on.","Hello! Thank you for calling Inchcape Customer Support. I'm here to help. I understand you're concerned because you were promised a discount that's not shown on your invoice. To assist you better, could you please provide me with the invoice code and date? This will help me validate the details and identify the issue."
-Calm,Payment,There are duplicate charges on my bank for the same service. I need this corrected immediately.,"Hello, thank you for calling Inchcape Customer Support. I'm here to help. I understand there are duplicate charges on your bank statement for the same service. I apologize for the inconvenience. May I have the invoice code and date of the service for verification? Once confirmed, I will work on correcting this immediately.
-
-(After validation)
-
-I see the duplicate charges now. I'm sorry for the trouble this has caused. I have corrected the issue, and you should see the refund in your account within 3-5 business days. May I recap by saying that the duplicate charges have been addressed, and a refund is on its way? Is there anything else I can assist you with today?"
-Calm,Payment,"It’s been over a week since my service, and I still haven’t received my receipt. When will I get it?","Hello! Thank you for calling Inchcape Customer Support, where we strive for distribution excellence. I'm [Your Name]. I understand you haven't received your receipt a week after your service. Let me look into this for you.
-
-To better assist you, can you please provide me with your invoice details, such as the invoice code and date? This will help me validate the information and identify the issue.
-
-While I investigate, I would like to remind you that our call is being recorded for safety and quality assurance purposes. I appreciate your patience, and I will provide a clear solution as soon as possible."
-Calm,Payment, The descriptions don’t match the services I received. Can you clarify this for me?,"Hello, I understand your concern about the discrepancy between the services you received and the descriptions provided. I'm here to help clarify this for you. May you please tell me more about the issue and the specific descriptions that are causing confusion? This will help me better understand and provide an accurate solution for you."
-Calm,Warranty,"I brought my car in for a repair, but they told me the warranty won’t cover it. This isn’t what I was told when I bought the warranty!","Hello, I'm sorry to hear about your experience. I understand how frustrating this must be. To assist you better, could you please provide me with the warranty code? I'll review the warranty terms and limitations, including the duration of coverage and conditions that may void the warranty. I'll then provide a resolution for you."
-Calm,Warranty,"I paid extra for the extended warranty, but now I’m being charged for repairs that should be covered. What’s the point of having this warranty?","Hello, I understand how frustrating this must be. I'd be happy to look into this for you. Could you please provide me with the warranty code and the details of the repairs you're being charged for? Let's see if they're covered under your EXT-02 Extended Warranty."
-Calm,Warranty,"I submitted my claim weeks ago, and I still haven’t heard anything back. How long is this supposed to take?","Hello, I understand your concern about not hearing back on your claim. Typically, our claims process takes up to two weeks. However, I apologize for the delay in your case. I will escalate this to the appropriate department to get an update for you. Thank you for your patience."
-Calm,Warranty,"When I bought the car, they told me the warranty would cover everything. Now, I’m finding out that certain parts aren’t covered. This feels misleading.","I understand how frustrating this must feel. Let's validate your warranty code to review the terms. The warranty for your car covers basic parts and labor for up to 3 years or 36,000 miles (STD-01). However, there might be specific parts with separate coverage, such as the battery, alternator, brake, and headlights, which can be covered for up to 2 years, regardless of mileage (PRT-03). I'm here to help, so let's clarify which parts you're concerned about, and I'll provide the best resolution possible."
-Angry,Payment,"I was promised a damn discount on my service, but the invoice doesn’t reflect it. What the hell is going on here?","Hello, thank you for calling Inchcape Customer Support. I apologize for the inconvenience you're experiencing. I'd like to understand more about the issue. Can you tell me the invoice code and date, as well as the details of the promised discount? I'm here to help and will do my best to resolve this for you. Thank you for your patience."
-Angry,Payment,"Why the hell are there duplicate charges on my bank account for the same service? This better get corrected immediately, or I’m gonna lose my shit.","Hello, I understand how frustrating this must be. I appreciate your patience while I validate the invoice details and investigate these duplicate charges. I will work to resolve this issue promptly for you."
-Angry,Payment,"It’s been over a week since my service, and I still haven’t gotten my damn receipt! When the fuck am I supposed to get it?","Hello, I understand how frustrating this must be. I apologize for the inconvenience you've experienced with your receipt. Let me look into this for you. May I have your invoice code and date, please? This will help me validate the details and provide a resolution more efficiently. Thank you for your patience."
-Angry,Payment,"The descriptions on this invoice don’t match a single thing I actually had done. Can you explain this shit, or what?","I understand how frustrating this must be. I'd be happy to help you validate the invoice details. Could you please provide me with the invoice code, date, and a brief description of the services listed? This will allow me to better understand and address your concerns."
-Angry,Warranty,"I brought my car in for a repair, and now you’re telling me the fucking warranty won’t cover it? That’s not what I was told when I bought the damn thing!","I'm really sorry for the confusion and any frustration this has caused you. Let's validate the warranty code for your vehicle and review the terms together to clarify the coverage. I understand how important this is to you, and I'm here to help."
-Angry,Warranty,"I paid extra for the extended warranty, and now you’re charging me for repairs that should be covered? What the hell is the point of even having this warranty?","Hello, I understand how frustrating this must be. Let me validate your warranty code and review the terms for you. The Extended Warranty (EXT-02) covers major components for up to 5 years or 60,000 miles. I'll check if the repairs fall under this warranty. Thank you for your patience."
-Angry,Warranty,"I submitted my claim weeks ago, and I still haven’t heard a damn thing back. How long is this supposed to take, or do you all just not care?","Hello, I understand your frustration with not receiving an update on your claim. I apologize for any inconvenience this may have caused. Typically, claim processing takes some time, but I will look into this for you and provide an update as soon as possible. Thank you for your patience. Is there anything else I can assist you with today?"
-Angry,Warranty,"When I bought the car, they told me the warranty would cover everything. Now, suddenly, parts aren’t covered? This is completely misleading and a load of crap!","I'm really sorry to hear about your frustration. I understand how misleading this situation might seem. Let's validate your warranty code and review the terms together to clarify the coverage. This way, we can ensure that we're providing you with the correct information."
